requirements.txt
# API Clients
twitchio==2.8.2
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Configuration
pyyaml==6.0.1

# HTTP Requests
requests==2.31.0
aiohttp==3.9.1

# Web Framework
flask==3.0.0

# Database
sqlalchemy==2.0.39